0.2 release: Changed the folder structure a little, added motion control for the switch, added 5 tracks

+++ include/ColorUtility.h
@@ -4,7 +4,7 @@
 #include <SDL2/SDL.h>
 
 SDL_Color RandomWallColor();
-SDL_Color GetNextColor(SDL_Color current, SDL_Color target, int speed);
+SDL_Color GetNextColor(SDL_Color current, SDL_Color target, uint8_t speed);
 SDL_Color RGBtoGBR(SDL_Color origin);
 SDL_Color MultiplyColor(SDL_Color c, float f);
 SDL_Color Negative(SDL_Color c);
